Make event_handler a config function
