+++ data_integration/logging/events.py
@@ -1,27 +1,11 @@
 """Events that are emitted during pipeline execution"""
 
-import abc
 import datetime
 import json
 
 import enum
 
-
-class Event():
-    def __init__(self) -> None:
-        """
-        Base class for events that are emitted from mara.
-        """
-
-    def to_json(self):
-        return json.dumps({field: value.isoformat() if isinstance(value, datetime.datetime) else value
-                           for field, value in self.__dict__.items()})
-
-
-class EventHandler(abc.ABC):
-    @abc.abstractmethod
-    def handle_event(self, event: Event):
-        pass
+from ..event_base import Event
 
 
 class PipelineEvent(Event):
