+++ data_integration/logging/run_log.py
@@ -7,6 +7,7 @@ from sqlalchemy.ext.declarative import declarative_base
 import mara_db.postgresql
 from .. import config
 from ..logging import events, system_statistics
+from .. import event_base
 
 Base = declarative_base()
 
@@ -85,11 +86,11 @@ WHERE run_id = {"%s"} and end_time IS NULL
         cursor.execute(_close_run, (run_id,))
 
 
-class RunLogger(events.EventHandler):
+class RunLogger(event_base.EventHandler):
     run_id: int = None
     node_output: {tuple: [events.Output]} = None
 
-    def handle_event(self, event: events.Event):
+    def handle_event(self, event: event_base.Event):
 
         if isinstance(event, events.RunStarted):
             with mara_db.postgresql.postgres_cursor_context('mara') as cursor:  # type: psycopg2.extensions.cursor
@@ -120,7 +121,7 @@ RETURNING node_run_id''', (self.run_id, event.node_path, event.start_time, event
         elif isinstance(event, system_statistics.SystemStatistics):
             with mara_db.postgresql.postgres_cursor_context('mara') as cursor:  # type: psycopg2.extensions.cursor
                 cursor.execute(f'''
-INSERT INTO data_integration_system_statistics (timestamp, disc_read, disc_write, net_recv, net_sent, 
+INSERT INTO data_integration_system_statistics (timestamp, disc_read, disc_write, net_recv, net_sent,
                                   cpu_usage, mem_usage, swap_usage, iowait)
 VALUES ({"%s, %s, %s, %s, %s, %s, %s, %s, %s"})''',
                                (event.timestamp, event.disc_read, event.disc_write, event.net_recv,
@@ -136,7 +137,7 @@ RETURNING node_run_id''', (event.end_time, event.succeeded, self.run_id, event.n
                 node_run_id = cursor.fetchone()[0]
 
                 cursor.execute('''
-INSERT INTO data_integration_node_output (node_run_id, timestamp, message, format, is_error) 
+INSERT INTO data_integration_node_output (node_run_id, timestamp, message, format, is_error)
 VALUES ''' + ','.join([cursor.mogrify('(%s,%s,%s,%s,%s)', (node_run_id, output_event.timestamp, output_event.message,
                                                            output_event.format, output_event.is_error))
                       .decode('utf-8')
@@ -145,7 +146,7 @@ VALUES ''' + ','.join([cursor.mogrify('(%s,%s,%s,%s,%s)', (node_run_id, output_e
         elif isinstance(event, events.RunFinished):
             with mara_db.postgresql.postgres_cursor_context('mara') as cursor:  # type: psycopg2.extensions.cursor
                 cursor.execute(f'''
-UPDATE data_integration_run 
+UPDATE data_integration_run
 SET end_time={"%s"}, succeeded={"%s"}
 WHERE run_id={"%s"}''', (event.end_time, event.succeeded, self.run_id))
 
@@ -157,11 +158,11 @@ DELETE FROM data_integration_node_output WHERE node_run_id IN (
 
                 cursor.execute(f'''
 DELETE FROM data_integration_node_run WHERE run_id IN (
-    SELECT run_id FROM data_integration_run 
+    SELECT run_id FROM data_integration_run
     WHERE start_time + INTERVAL '{config.run_log_retention_in_days()} days' < current_timestamp);''')
 
                 cursor.execute(f'''
-DELETE FROM data_integration_run 
+DELETE FROM data_integration_run
 WHERE start_time + INTERVAL '{config.run_log_retention_in_days()} days' < current_timestamp;''')
 
                 cursor.execute(f'''
